+++ crates/fuzz/src/traits/transaction.rs
@@ -50,9 +50,6 @@ pub trait TransactionMethods:
         if fuzzing_metrics.is_ok() {
             let mut stats_logger = FuzzingStatistics::new();
 
-            // Record transaction invocation
-            stats_logger.increase_invoked(self.get_transaction_name());
-
             // Run pre-transaction hook
             self.pre_transaction(client);
 
@@ -60,12 +57,16 @@ pub trait TransactionMethods:
             self.set_snapshot_before(client);
 
             // Execute the transaction
-            let tx_result = client.process_instructions(&instructions);
+            client._increment_transaction_execution(self.get_transaction_name(), &mut stats_logger);
+            let tx_result = client._process_instructions(&instructions);
 
             match tx_result {
                 Ok(_) => {
                     // Record successful execution
-                    stats_logger.increase_successful(self.get_transaction_name());
+                    client._increment_transaction_success(
+                        self.get_transaction_name(),
+                        &mut stats_logger,
+                    );
 
                     // Take snapshot of accounts after execution
                     self.set_snapshot_after(client);
@@ -90,8 +91,11 @@ pub trait TransactionMethods:
                     self.post_transaction(client);
                 }
                 Err(e) => {
-                    // Record transaction failure
-                    stats_logger.increase_failed(self.get_transaction_name());
+                    client._record_transaction_error(
+                        self.get_transaction_name(),
+                        e.to_string(),
+                        &mut stats_logger,
+                    );
                     stats_logger.output_serialized();
 
                     // Handle transaction error
@@ -106,7 +110,7 @@ pub trait TransactionMethods:
             self.pre_transaction(client);
 
             // Execute the transaction
-            let tx_result = client.process_instructions(&instructions);
+            let tx_result = client._process_instructions(&instructions);
 
             match tx_result {
                 Ok(_) => {
@@ -153,29 +157,31 @@ pub trait TransactionMethods:
         if fuzzing_metrics.is_ok() {
             let mut stats_logger = FuzzingStatistics::new();
 
-            // Record transaction invocation
-            stats_logger.increase_invoked(self.get_transaction_name());
-
             // Take snapshot of accounts before execution
             self.set_snapshot_before(client);
 
+            client._increment_transaction_execution(self.get_transaction_name(), &mut stats_logger);
             // Execute the transaction
-            let tx_result = client.process_instructions(&instructions);
+            let tx_result = client._process_instructions(&instructions);
 
             match tx_result {
                 Ok(_) => {
-                    // Record successful execution
-                    stats_logger.increase_successful(self.get_transaction_name());
-
+                    client._increment_transaction_success(
+                        self.get_transaction_name(),
+                        &mut stats_logger,
+                    );
                     // Output statistics
                     stats_logger.output_serialized();
 
                     // Take snapshot of accounts after execution
                     self.set_snapshot_after(client);
                 }
-                Err(_e) => {
-                    // Record transaction failure
-                    stats_logger.increase_failed(self.get_transaction_name());
+                Err(e) => {
+                    client._record_transaction_error(
+                        self.get_transaction_name(),
+                        e.to_string(),
+                        &mut stats_logger,
+                    );
                     stats_logger.output_serialized();
                 }
             }
@@ -184,7 +190,7 @@ pub trait TransactionMethods:
             self.set_snapshot_before(client);
 
             // Execute the transaction
-            let tx_result = client.process_instructions(&instructions);
+            let tx_result = client._process_instructions(&instructions);
 
             match tx_result {
                 Ok(_) => {